python_scoring/requirements.txt
fastapi==0.109.0
uvicorn==0.23.0
joblib==1.3.2
pandas==2.2.2
numpy==1.26.2
pydantic==2.6.2
python-multipart==0.0.6
scikit-learn==1.7.2
loguru==0.7.0
mysql-connector-python==9.5.0
python-dotenv==1.0.1
